devops-credentials-api/requirements.txt
flask==1.1.2
py-healthcheck
elasticsearch>=7.7.1,<8.0.0
requests>=2.23.1